Add dto validations, add http exception when product was not found

+++ src/product/service/product.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { CreateProductDto } from "../dto/create-product.dto";
 import { UpdateProductDto } from "../dto/update-product.dto";
 import { ProductRepository } from "../repository/product.repository";
@@ -17,7 +17,13 @@ export class ProductService {
   }
 
   async findOne(id: number): Promise<Product> {
-    return await this.productRepository.findOne(id);
+    const product = await this.productRepository.findOne(id);
+
+    if (!product) {
+      throw new NotFoundException("Product not found");
+    }
+
+    return product;
   }
 
   async create(createProductDto: CreateProductDto): Promise<Product> {
@@ -25,12 +31,27 @@ export class ProductService {
     return await this.productRepository.create(mappedProduct);
   }
 
-  async update(id: number, updateProductDto: UpdateProductDto): Promise<Product>  {
+  async update(
+    id: number,
+    updateProductDto: UpdateProductDto
+  ): Promise<Product> {
+    const product = await this.productRepository.findOne(id);
+
+    if (!product) {
+      throw new NotFoundException("Product not found");
+    }
+
     const mappedProduct = fromUpdatedDtoToEntity(updateProductDto);
     return await this.productRepository.update(id, mappedProduct);
   }
 
   async remove(id: number): Promise<void> {
+    const product = await this.productRepository.findOne(id);
+
+    if (!product) {
+      throw new NotFoundException("Product not found");
+    }
+
     return await this.productRepository.remove(id);
   }
 }
